Test React components and CMS after linking
This provides some insurance that projects still work after linking.

We use decicated Cypress GitHub Action for running tests. This allows us to run
the server and Cypress in separate threads.

We only run a single Cypress test for React components. That should be
sufficient to ensure that the project is still intact.