Ensure that Docker Compose within projects work as expected
When setting up projects configure Dapple-specific COMPOSE_PROJECT_NAME 
values in the project .env files. This will ensure that containers will
not conflict with default project checkouts using defaults and that 
subsequent docker commands will work on the Dapple-initiated containers.

We use variables instead of environment variables as they are easier to
work with. Their actual values are shown when running commands.

Update our test to ensure that it works as expected.