Only install dependencies as a part of initialization when needed
This ensures that we do not keep any docker containers running for
developers who do not want/use them.

Sources/generates ensures that nothing is installed needlessly.

Instead we run dev:reset when starting cms development. This also means
that we have to move cms:link from a dependecy to a command since we
cannot ensure that we have a running CMS instance beforehand.

This also means that we have to start the CMS for development explicitly
when running tests. We do not use dev:* commands here as they start
running watch processes. Instead we run their subtasks individually.

Since there is no reset anymore we cannot test it either.